Spent 5 or 6 hours tracking down an align bug in create object - Universial Motion section.   Damn it Jim, I am a high level programmer not a down in the bits and bytes weenie.... :)

+++ Source/ACE/Entity/WorldObject.cs
@@ -2432,7 +2432,8 @@ namespace ACE.Entity
                 {
                     CurrentMotionState = new UniversalMotion(MotionStance.Standing);
                     var movementData = CurrentMotionState.GetPayload(wo.Guid, wo.Sequences);
-                    writer.Write(movementData.Length);
+                    // May not need this cast from int to uint, but the protocol says uint Og II
+                    writer.Write((uint)movementData.Length);
                     writer.Write(movementData);
                     uint autonomous = CurrentMotionState.IsAutonomous ? (ushort)1 : (ushort)0;
                     writer.Write(autonomous);
